pyproject.toml
[project]
name = "fastapicloud"
version = "0.1.0"
description = "FastAPI Cloud Service Framework"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.8.1",
    "alembic>=1.15.1",
    "sqlalchemy[asyncio]>=2.0.0",
    "sqlmodel>=0.0.14",
    "redis>=5.0.0",
    "requests>=2.31.0",
    "httpx>=0.24.0",
    "typer>=0.15.2",
    "pyyaml>=6.0",
    "loguru>=0.7.0",
    "fastapi-pagination>=0.12.0",
    "PyJWT>=2.10.1",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.1",
    "python-multipart>=0.0.20",
    "user-agents>=2.2.0",
    "openai>=1.76.0",
    "python-jose[cryptography]>=3.3.0",
    "aiosqlite>=0.19.0",
    "Jinja2>=3.1.6",
    "psutil>=5.9.0"
]

[tool.setuptools.packages]
find = { include = ["app", "core"] }

[tool.uv]
index-url = "https://pypi.org/simple"
